[SYCL] Refactor program_impl
